review(pr1): revue de code feature/auth-api — corrections appliquées
Problèmes détectés et corrigés:
1. notes.service.js: validation préalable des notes AVANT BEGIN transaction
   - Évite ROLLBACK partiel sur lot de notes avec valeur invalide
   - La validation est maintenant atomique (tout ou rien)
2. messages.service.js: commentaire explicatif sur le 403 intentionnel
   - getMessageById retourne 403 (pas 404) pour ne pas révéler l'existence
   - Sécurité: empêche l'énumération des IDs de messages
3. auth.routes.js: ajout .toInt() sur etablissement_id
   - Forçage de la conversion entière pour éviter les valeurs flottantes
Points positifs: architecture modulaire, JWT correct, SQL paramétré, bcrypt salt 12
REVIEW_PR1.md: rapport complet de la revue

+++ backend/src/modules/auth/auth.routes.js
@@ -13,7 +13,7 @@ router.post(
   [
     body('email').isEmail().withMessage('Email invalide').normalizeEmail(),
     body('mot_de_passe').notEmpty().withMessage('Mot de passe requis'),
-    body('etablissement_id').isInt({ min: 1 }).withMessage('Établissement requis'),
+    body('etablissement_id').isInt({ min: 1 }).toInt().withMessage('Établissement requis'),
   ],
   validate,
   authController.login
